Move structures.py and restate import path
